Squashed commit of the following:
commit 9a9843d7a5b0110e6b0f7b4e4527d2f160750eec
Author: Sangmeeeee <goodman1223@naver.com>
Date:   Fri Aug 5 10:49:55 2022 +0900

    [fix] : 회원가입시 picture를 입력하지 않았을 경우 user info 요청시 발생하던 error 처리

commit c50b764276c2f2b879eefe6afeae3b598ac8edb2
Merge: f7f318d 075f65b
Author: 이기종 <rl00822@naver.com>
Date:   Thu Aug 4 15:06:01 2022 +0900

    Merge branch 'S07P12D206-139/feature/stt/request' into 'back-develop'

    S07P12D206-139 [feat] : STT 요청 기능 추가

    See merge request s07-webmobile1-sub2/S07P12D206!37

commit 075f65bb7061c6e43aa30027ba110f43da7155a4
Author: 이기종 <rl00822@naver.com>
Date:   Thu Aug 4 15:06:01 2022 +0900

    S07P12D206-139 [feat] : STT 요청 기능 추가

+++ back/demo/src/main/java/com/dearme/demo/domain/board/service/CommentServiceImpl.java
@@ -7,9 +7,15 @@ import com.dearme.demo.domain.board.dto.comment.CommentUpdateResponseDto;
 import com.dearme.demo.domain.board.entity.Board;
 import com.dearme.demo.domain.board.entity.Comment;
 import com.dearme.demo.domain.board.exception.board.NoExistBoardException;
+import com.dearme.demo.domain.board.exception.comment.NoCommentDeletePermissionException;
+import com.dearme.demo.domain.board.exception.comment.NoCommentSavePermissionException;
+import com.dearme.demo.domain.board.exception.comment.NoCommentUpdatePermissionException;
 import com.dearme.demo.domain.board.exception.comment.NoExistCommentException;
 import com.dearme.demo.domain.board.repository.BoardRepository;
 import com.dearme.demo.domain.board.repository.CommentRepository;
+import com.dearme.demo.domain.user.entity.Type;
+import com.dearme.demo.domain.user.entity.User;
+import com.dearme.demo.domain.user.exception.NoExistUserException;
 import com.dearme.demo.domain.user.repository.UserRepository;
 import com.dearme.demo.global.util.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
@@ -33,19 +39,17 @@ public class CommentServiceImpl implements CommentService{
         Board board = boardRepository.findBoardByBoardid(boardid).orElseThrow(()->{
             throw new NoExistBoardException();
         });
-        System.out.println(board);
-        System.out.println(comment);
-//        User user = userRepository.findUserById(id).orElseThrow(() -> {
-//            throw new NoExistUserException();
-//        });
-//        if(user.getType().equals("COUNSELOR")){
+        User user = userRepository.findUserById(id).orElseThrow(() -> {
+            throw new NoExistUserException();
+        });
+        if(user.getType().equals(Type.COUNSELOR)){
         comment.setBoard(board);
-//           comment.setUser(user);
+           comment.setUser(user);
             commentRepository.save(comment);
             return new CommentSaveResponseDto(comment.getCommentid());
-//        }else{
-//            throw new NoCommentSavePermissionException();
-//        }
+        }else{
+            throw new NoCommentSavePermissionException();
+        }
 
     }
 
@@ -54,15 +58,15 @@ public class CommentServiceImpl implements CommentService{
         Comment comment = commentRepository.findCommentByCommentid(commentid).orElseThrow(()->{
             throw new NoExistCommentException();
         });
-//        User user = userRepository.findUserById(id).orElseThrow(() -> {
-//            throw new NoExistUserException();
-//        });
-//        if(user.getUserId().equals(comment.getUser().getUserId())){
-            comment.update(dto.getContents(), dto.getDate());
+        User user = userRepository.findUserById(id).orElseThrow(() -> {
+            throw new NoExistUserException();
+        });
+        if(user.getUserId().equals(comment.getUser().getUserId())){
+            comment.update(dto.getDate(), dto.getContents());
             return new CommentUpdateResponseDto(comment.getCommentid());
-//        }else{
-//            throw new NoCommentUpdatePermissionException();
-//        }
+        }else{
+            throw new NoCommentUpdatePermissionException();
+        }
     }
 
     @Transactional
@@ -70,14 +74,14 @@ public class CommentServiceImpl implements CommentService{
         Comment comment = commentRepository.findCommentByCommentid(commentid).orElseThrow(()->{
             throw new NoExistCommentException();
         });
-//        User user = userRepository.findUserById(id).orElseThrow(() -> {
-//            throw new NoExistUserException();
-//        });
-//        if(user.getUserId().equals(comment.getUser().getUserId())){
+        User user = userRepository.findUserById(id).orElseThrow(() -> {
+            throw new NoExistUserException();
+        });
+        if(user.getUserId().equals(comment.getUser().getUserId())){
             commentRepository.delete(comment);
-//        }else{
-//            throw new NoCommentDeletePermissionException();
-//        }
+        }else{
+            throw new NoCommentDeletePermissionException();
+        }
     }
 
 }
